Conflicts:
	app-aux/build-static.pl
	autogen.sh

+++ autogen.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 #
-# autogen.sh - Generates initial makefiles from a pristine GIT tree
+# autogen.sh - Generates initial makefiles from a pristine CVS tree
 #
 # USAGE:
 #   autogen.sh [configure options]
@@ -12,10 +12,14 @@
 #  programs that would be run.
 #   e.g. DRYRUN=1 ./autogen.sh
 #
-# AUTOMAKE ACLOCAL AUTOCONF AUTOHEADER LIBTOOLIZE
+# AUTOMAKE ACLOCAL AUTOCONF AUTOHEADER LIBTOOLIZE GTKDOCIZE
 #  If set (named after program) then this overrides any searching for
 #  the programs on the current PATH.
 #   e.g. AUTOMAKE=automake-1.7 ACLOCAL=aclocal-1.7 ./autogen.sh
+#
+# CONFIG_DIR (default ../config)
+#  The directory where fresh GNU config.guess and config.sub can be
+#  found for automatic copying in-place.
 #
 # PATH
 #  Where the programs are searched for
@@ -24,7 +28,7 @@
 #  Source directory
 #
 # This script is based on similar scripts used in various tools
-# commonly made available via GIT and used with GNU automake.
+# commonly made available via CVS and used with GNU automake.
 # Try 'locate autogen.sh' on your system and see what you get.
 #
 # This script is in the public domain
@@ -33,19 +37,24 @@
 # Directory for the sources
 SRCDIR=${SRCDIR-.}
 
-# Because GIT doesn't support empty directories
-if [ ! -d "$SRCDIR/build-scripts" ]; then
-    mkdir "$SRCDIR/build-scripts"
-fi
+# Where the GNU config.sub, config.guess might be found
+CONFIG_DIR=${CONFIG_DIR-../config}
 
 # The programs required for configuring which will be searched for
 # in the current PATH.
 # Set an envariable of the same name in uppercase, to override scan
 #
 programs="automake aclocal autoconf autoheader libtoolize"
+confs=`find . -name configure.ac -print`
+if grep "^AC_CHECK_PROGS.SWIG" $confs >/dev/null; then
+  programs="$programs swig"
+fi
+ltdl=
+if grep "^AC_LIBLTDL_" $confs >/dev/null; then
+  ltdl="--ltdl"
+fi
 
 # Some dependencies for autotools:
-# automake 1.11 requires autoconf 2.62
 # automake 1.10 requires autoconf 2.60
 # automake 1.9 requires autoconf 2.58
 # automake 1.8 requires autoconf 2.58
@@ -61,8 +70,12 @@ swig_min_vers=010324
 automake_args="--add-missing"
 autoconf_args=
 libtoolize_args="$ltdl --force --copy --automake"
-aclocal_args="-I /usr/local/share/aclocal"
+aclocal_args=""
+if test -d /usr/local/share/aclocal; then
+  aclocal_args="-I /usr/local/share/aclocal"
+fi
 automake_args="--gnu --add-missing --force --copy"
+
 
 # You should not need to edit below here
 ######################################################################
@@ -151,7 +164,7 @@ update_prog_version() {
 
   nameglob="$prog*"
   if [ -x /usr/bin/uname ]; then
-    if [ `/usr/bin/uname`x = 'Darwinx' -a $prog = 'libtoolize' ] ; then
+    if [ `/usr/bin/uname` = 'Darwin' -a $prog = 'libtoolize' ] ; then
       nameglob="g$nameglob"
     fi
   fi
@@ -246,39 +259,62 @@ done
 
 echo "$program: Dependencies satisfied"
 
-cd "$SRCDIR"
-# Ensure that these are created by the versions on this system
-# (indirectly via automake)
-$DRYRUN rm -f ltconfig ltmain.sh libtool stamp-h*
-# Made by automake
-$DRYRUN rm -f missing depcomp
-# automake junk
-$DRYRUN rm -rf autom4te*.cache
-
-config_macro_dir=`sed -ne 's/^AC_CONFIG_MACRO_DIR(\([^)]*\).*/\1/p' configure.ac`
-if test "X$config_macro_dir" = X; then
-  config_macro_dir=.
-else
-  aclocal_args="$aclocal_args -I $config_macro_dir "
+if test -d $SRCDIR/libltdl; then
+  touch $SRCDIR/libltdl/NO-AUTO-GEN
 fi
 
-config_aux_dir=`sed -ne 's/^AC_CONFIG_AUX_DIR(\([^)]*\).*/\1/p' configure.ac`
-if test "X$config_aux_dir" = X; then
-  config_aux_dir=.
+config_dir=
+if test -d $CONFIG_DIR; then
+  config_dir=`cd $CONFIG_DIR; pwd`
 fi
 
-echo "$program: Running $libtoolize $libtoolize_args"
-$DRYRUN rm -f ltmain.sh libtool
-eval $DRYRUN $libtoolize $libtoolize_args
 
-echo "$program: Running $aclocal $aclocal_args"
-$DRYRUN $aclocal $aclocal_args
-echo "$program: Running $autoheader"
-$DRYRUN $autoheader
-echo "$program: Running $automake $automake_args"
-$DRYRUN $automake $automake_args
-echo "$program: Running $autoconf"
-$DRYRUN $autoconf $autoconf_args
+for coin in `find $SRCDIR -name configure.ac -print`
+do 
+  dir=`dirname $coin`
+  if test -f "$dir/NO-AUTO-GEN"; then
+    echo $program: Skipping $dir -- flagged as no auto-gen
+  else
+    echo " "
+    echo $program: Processing directory $dir
+    ( cd "$dir"
+
+      # Ensure that these are created by the versions on this system
+      # (indirectly via automake)
+      $DRYRUN rm -f ltconfig ltmain.sh libtool stamp-h*
+      # Made by automake
+      $DRYRUN rm -f missing depcomp
+      # automake junk
+      $DRYRUN rm -rf autom4te*.cache
+
+      if test "X$config_dir" != X; then
+        echo "$program: Updating config.guess and config.sub"
+	for file in config.guess config.sub; do
+	  cfile=$config_dir/$file
+	  if test -f $cfile; then
+	    $DRYRUN rm -f $file
+	    $DRYRUN cp -p $cfile $file
+	  fi
+	done
+      fi
+
+      echo "$program: Running $libtoolize $libtoolize_args"
+      $DRYRUN rm -f ltmain.sh libtool
+      eval $DRYRUN $libtoolize $libtoolize_args
+
+      echo "$program: Running $aclocal $aclocal_args"
+      $DRYRUN $aclocal $aclocal_args
+      if grep "^AM_CONFIG_HEADER" configure.ac >/dev/null; then
+	echo "$program: Running $autoheader"
+	$DRYRUN $autoheader
+      fi
+      echo "$program: Running $automake $automake_args"
+      $DRYRUN $automake $automake_args $automake_args
+      echo "$program: Running $autoconf"
+      $DRYRUN $autoconf $autoconf_args
+    )
+  fi
+done
 
 
 rm -f config.cache
